Gigant refactor of project structure to ease variations
Fixed crlf issue
