+++ include/common/Patterns.h
