+++ include/optimized/GameOfLife.h
